requirements.txt
asgiref==3.8.1
Django==5.2
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-js-asset==3.1.2
pillow==11.2.1
sqlparse==0.5.3
